displaying photog credit on each item
